# Conflicts:
#	dune/stuff/aliases.hh
#	dune/stuff/common/algorithm.hh
#	dune/stuff/common/configuration.cc
#	dune/stuff/common/configuration.hh
#	dune/stuff/common/debug.hh
#	dune/stuff/common/float_cmp.hh
#	dune/stuff/common/fvector.hh
#	dune/stuff/common/math.hh
#	dune/stuff/common/memory.hh
#	dune/stuff/common/parallel/threadmanager.cc
#	dune/stuff/common/profiler.cc
#	dune/stuff/common/profiler.hh
#	dune/stuff/common/random.hh
#	dune/stuff/common/ranges.hh
#	dune/stuff/common/string.hh
#	dune/stuff/common/tuple.hh
#	dune/stuff/common/type_utils.hh
#	dune/stuff/common/vector.hh
#	dune/stuff/functions/ESV2007.hh
#	dune/stuff/functions/constant.hh
#	dune/stuff/functions/expression.hh
#	dune/stuff/functions/expression/mathexpr.cc
#	dune/stuff/functions/flattop.hh
#	dune/stuff/functions/interfaces.hh
#	dune/stuff/functions/spe10model2.hh
#	dune/stuff/grid/boundaryinfo.hh
#	dune/stuff/grid/information.hh
#	dune/stuff/grid/intersection.hh
#	dune/stuff/grid/layers.hh
#	dune/stuff/grid/output/entity_visualization.hh
#	dune/stuff/grid/output/pgf.hh
#	dune/stuff/grid/periodicview.hh
#	dune/stuff/grid/provider.hh
#	dune/stuff/grid/provider/cube.hh
#	dune/stuff/grid/provider/default.hh
#	dune/stuff/grid/provider/dgf.hh
#	dune/stuff/grid/provider/eoc.hh
#	dune/stuff/grid/provider/gmsh.hh
#	dune/stuff/grid/provider/interface.hh
#	dune/stuff/grid/search.hh
#	dune/stuff/grid/structuredgridfactory.hh
#	dune/stuff/grid/walker.hh
#	dune/stuff/la/container.hh
#	dune/stuff/la/container/common.hh
#	dune/stuff/la/container/container-interface.hh
#	dune/stuff/la/container/eigen/base.hh
#	dune/stuff/la/container/eigen/dense.hh
#	dune/stuff/la/container/eigen/sparse.hh
#	dune/stuff/la/container/istl.hh
#	dune/stuff/la/container/matrix-interface.hh
#	dune/stuff/la/container/vector-interface.hh
#	dune/stuff/la/solver/common.hh
#	dune/stuff/la/solver/eigen.hh
#	dune/stuff/la/solver/istl.hh
#	dune/stuff/la/solver/istl_amg.hh
#	dune/stuff/playground/functions/ESV2007.hh
#	dune/stuff/playground/grid/provider/cornerpoint.hh
#	dune/stuff/test/CMakeLists.txt
#	dune/stuff/test/grid_provider.hh
#	dune/stuff/test/grid_provider_cube.cc

+++ cmake/modules/DuneStuffMacros.cmake
@@ -33,6 +33,25 @@ CHECK_CXX_SOURCE_COMPILES("
 "  HAS_STD_BEGIN_END
 )
 
+CHECK_CXX_SOURCE_COMPILES("
+    int main(void)
+    {
+      int a __attribute__((unused)) = 0;
+    };
+"  HAS_WORKING_UNUSED_ATTRIBUTE
+)
+
+CHECK_CXX_SOURCE_COMPILES("
+    #include <map>
+    int main(void)
+    {
+      std::map<int, int> a;
+      a.emplace(2, 2);
+      return 0;
+    };
+"  HAVE_MAP_EMPLACE
+)
+
 Macro(ADD_IF_SUPPORTED dest)
   FOREACH(flag ${ARGN})
     CHECK_CXX_ACCEPTS_FLAG("${flag}" has_${flag})
@@ -152,7 +171,7 @@ macro(add_analyze)
 endmacro(add_analyze)
 
 macro(add_format)
-    find_program(FORMAT NAMES clang-format clang-format-3.4 clang-format-3.5 clang-format-3.6 clang-format-3.7 clang-format-3.8)
+    find_program(FORMAT NAMES clang-format clang-format-3.4 clang-format-3.5 clang-format-3.6 clang-format-3.7 clang-format-3.8 clang-format-3.9)
     if(EXISTS ${FORMAT})
         message(STATUS "adding format target")
         add_custom_target( format SOURCES ${ARGN} )
@@ -172,6 +191,8 @@ endmacro(add_format)
 
 macro(add_forced_doxygen_target)
   add_doxygen_target()
-  add_custom_target(doxygen_${ProjectName}_pre_build COMMAND rm -rf ${CMAKE_CURRENT_BINARY_DIR}/html )
-  add_dependencies(doxygen_${ProjectName} doxygen_${ProjectName}_pre_build)
+  if(TARGET doxygen_${ProjectName})
+    add_custom_target(doxygen_${ProjectName}_pre_build COMMAND rm -rf ${CMAKE_CURRENT_BINARY_DIR}/html )
+    add_dependencies(doxygen_${ProjectName} doxygen_${ProjectName}_pre_build)
+  endif()
 endmacro(add_forced_doxygen_target)
